refactor user model with domain module
